refactor(console): use single text input in sdk guide on app newly created
