--- packages/console/src/mdx-components/MultiTextInputField/index.module.scss
@@ -1,10 +1,0 @@
-@use '@/scss/underscore' as _;
-
-.field {
-  width: 556px;
-}
-
-.wrapper {
-  display: flex;
-  align-items: flex-start;
-}
