+++ packages/console/src/mdx-components/UriInputField/index.module.scss
@@ -1,0 +1,17 @@
+@use '@/scss/underscore' as _;
+
+.wrapper {
+  display: flex;
+  align-items: flex-start;
+  position: relative;
+}
+
+.field {
+  width: 556px;
+}
+
+.saveButton {
+  position: absolute;
+  left: calc(556px + _.unit(3));
+  top: 0;
+}
